various changes
