+++ src/Genimage/genimage-src/configure.ac
@@ -109,7 +109,5 @@ AC_CONFIG_HEADERS(config.h)
 AC_CONFIG_FILES([
         Makefile
 ])
-
 AC_REQUIRE_AUX_FILE([tap-driver.sh])
-
 AC_OUTPUT
